almost finished texture and mesh importers
mesh and textures are imported separatedly
