+++ MyGameEditor/imgui.ini
@@ -9,7 +9,6 @@ Size=400,419
 Collapsed=1
 
 [Window][Hierarchy]
-
 Pos=0,19
 Size=231,881
 Collapsed=0
@@ -22,7 +21,6 @@ Collapsed=0
 DockId=0x00000006,1
 
 [Window][Inspector]
-
 Pos=1354,19
 Size=246,881
 Collapsed=0
@@ -40,7 +38,6 @@ Collapsed=0
 DockId=0x00000006,1
 
 [Window][Configuration]
-
 Pos=233,662
 Size=1119,238
 Collapsed=0
@@ -52,6 +49,6 @@ DockSpace       ID=0x8B93E3BD Window=0xA787BDB4 Pos=0,19 Size=1600,881 Split=X
   DockNode      ID=0x00000005 Parent=0x8B93E3BD SizeRef=1367,881 Split=X
     DockNode    ID=0x00000002 Parent=0x00000005 SizeRef=1352,881 Split=Y
       DockNode  ID=0x00000004 Parent=0x00000002 SizeRef=1600,641 CentralNode=1
-      DockNode  ID=0x00000006 Parent=0x00000002 SizeRef=1600,238 Selected=0x49278EEE
+      DockNode  ID=0x00000006 Parent=0x00000002 SizeRef=1600,238 Selected=0x09A1ED00
     DockNode    ID=0x00000007 Parent=0x00000005 SizeRef=246,881 Selected=0xE7039252
