+++ MyGameEngine/GameObject.cpp
@@ -4,13 +4,9 @@ using namespace std;
 
 GameObject::GameObject(string meshPath = "", string texturePath = "")
 {
-	Component* transform = createComponent(TRANSFORM);
-	Component* mesh = createComponent(MESH, meshPath);
-	//Component* texture = createComponent(TEXTURE, texturePath);
-
-	addComponent(transform);
-	addComponent(mesh);
-	//addComponent(texture);
+	createComponent(TRANSFORM);
+	if (meshPath != "") createComponent(MESH, meshPath);
+	if (texturePath != "") createComponent(TEXTURE, texturePath);
 
 	this->name = "defaultName";
 }
@@ -19,37 +15,39 @@ GameObject::~GameObject()
 {
 }
 
-Component* GameObject::createComponent(component_type type, string path)
+void GameObject::createComponent(component_type type, string meshPath, string textPath)
 {
 	Component* newComponent = nullptr;
+	vector<Mesh::Ptr> mesh_shrdptrs;
 
+	stringstream ssmesh;
+	ssmesh << "..\\\\MyGameEditor\\\\Assets\\\\" << meshPath;
 	switch (type)
 	{
 	case TRANSFORM:
-		newComponent = new Transform();
+		newComponent = new TransformComp();
 		break;
 
 	case MESH:
-		newComponent = new Mesh(path);
+		// Check if files exist before loading
+		if (textPath != "") mesh_shrdptrs = Mesh::loadFromFile(meshPath, textPath);
+		else mesh_shrdptrs = Mesh::loadFromFile(ssmesh.str());
+
+		for (auto item : mesh_shrdptrs) {
+			newComponent = new MeshComp(item, ssmesh.str());
+		}
 		break;
 
 	case TEXTURE:
-
+		//newComponent = new TextureComp();
+		//component_vector.push_back(newComponent);
 		break;
 	}
-
-	return newComponent;
-}
-void GameObject::addComponent(Component* component)
-{
-	component_vector.push_back(component);
+	mesh_shrdptrs.clear();
 }
 
 void GameObject::drawObj()
 {
 	// Draws all the components in the vector of an object
-	for (Component* item : component_vector)
-	{
-		if (item->getActive()) item->drawComponent();
-	}
+
 }
