+++ MyGameEngine/GameObject.h
@@ -3,11 +3,12 @@
 
 #include "types.h"
 #include "Component.h"
-#include "Transform.h"
-#include "Mesh.h"
-
+#include "TransformComp.h"
+#include "MeshComp.h"
+#include "TextureComp.h"
 
 #include <string>
+#include <sstream>
 #include <vector>
 
 using namespace std;
@@ -19,9 +20,8 @@ public:
 	GameObject(string meshPath, string texturePath);
 	~GameObject();
 
-	Component* createComponent(component_type type, string path = "");
-	void addComponent(Component* component);
-	
+	void createComponent(component_type type, string meshPath = "", string textPath = "");
+
 	vector<Component*> component_vector;
 
 	void drawObj();
@@ -34,6 +34,7 @@ public:
 	{
 		this->name = name;
 	}
+	string adaptPath(string ogPath);
 
 private:
 	string name = "";
