+++ MyGameEngine/MeshComp.h
@@ -1,0 +1,48 @@
+#pragma once
+#include "Component.h"
+#include "Globals.h"
+#include "Mesh.h"
+
+#include <memory>
+#include <string>
+#include <vector>
+
+using namespace std;
+
+class MeshComp : public Component
+{
+public:
+	MeshComp(Mesh::Ptr mesh, string pathToSetName);
+	virtual ~MeshComp(){}
+
+	void Enable() { active = true; }
+	update_statusE Update() { return UPDATE_CONTINUEE; }
+	void Disable() { active = false; }
+
+	void extractName(string path);
+
+	void setMesh(Mesh::Ptr mesh)
+	{
+		_mesh = mesh;
+	}
+
+	string getName()
+	{
+		return this->name;
+	}
+
+	Mesh::Ptr getMeshData()
+	{
+		return _mesh;
+	}
+
+private:
+	Mesh::Ptr _mesh;
+	
+	component_type type = MESH;
+
+	string extension = "\.fbx";
+	string name = "";
+
+	bool active = true;
+};
