+++ MyGameEngine/MyGameEngine.vcxproj
@@ -23,15 +23,15 @@
     <ClInclude Include="Globals.h" />
     <ClInclude Include="Graphic.h" />
     <ClInclude Include="GraphicObject.h" />
+    <ClInclude Include="MeshComp.h" />
     <ClInclude Include="Mesh.h" />
-    <ClInclude Include="MeshImporter.h" />
     <ClInclude Include="EngineModule.h" />
     <ClInclude Include="Scene.h" />
+    <ClInclude Include="TextureComp.h" />
     <ClInclude Include="Texture.h" />
-    <ClInclude Include="Transform.h" />
+    <ClInclude Include="TransformComp.h" />
     <ClInclude Include="WonderEngine.h" />
     <ClInclude Include="Renderer.h" />
-    <ClInclude Include="Texture2D.h" />
     <ClInclude Include="types.h" />
   </ItemGroup>
   <ItemGroup>
@@ -43,15 +43,15 @@
     <ClCompile Include="CubeVertexBuffer.cpp" />
     <ClCompile Include="CubeWireframeIVBO.cpp" />
     <ClCompile Include="GraphicObject.cpp" />
+    <ClCompile Include="MeshComp.cpp" />
     <ClCompile Include="Mesh.cpp" />
-    <ClCompile Include="MeshImporter.cpp" />
     <ClCompile Include="GameObject.cpp" />
     <ClCompile Include="Scene.cpp" />
+    <ClCompile Include="TextureComp.cpp" />
     <ClCompile Include="Texture.cpp" />
-    <ClCompile Include="Transform.cpp" />
+    <ClCompile Include="TransformComp.cpp" />
     <ClCompile Include="WonderEngine.cpp" />
     <ClCompile Include="Renderer.cpp" />
-    <ClCompile Include="Texture2D.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <VCProjectVersion>17.0</VCProjectVersion>
@@ -106,7 +106,7 @@
 $(SolutionDir)\vcpkg\vcpkg.exe install opengl glew glm devil[libjpeg,libpng] assimp --recurse --triplet x64-windows</Command>
     </PreBuildEvent>
     <Lib>
-      <AdditionalDependencies>OpenGL32.lib;GLU32.lib;DevIL.lib;assimp-vc143-mt.lib</AdditionalDependencies>
+      <AdditionalDependencies>OpenGL32.lib;GLU32.lib;DevIL.lib;assimp-vc142-mt.lib</AdditionalDependencies>
     </Lib>
     <Lib>
       <AdditionalLibraryDirectories>$(SolutionDir)vcpkg\installed\x64-windows\lib\</AdditionalLibraryDirectories>
